Dependency fixes <version>jsk.17.6.5</version> next: <version>jsk.17.6.6</version>

+++ g-cluster/g-cluster-master/g-cluster-master-core/src/main/java/jsk/gcl/srv/logic/scaling/storage/GclNodeStorage.java
@@ -38,5 +38,5 @@ public interface GclNodeStorage extends RdbTransactionManager {
 
     List<GclNode> getInactiveNodes();
 
-    long getActiveNodeCount();
+    int getActiveNodeCount();
 }
